@@ -1,8 +1,3 @@
 # Be sure to restart your server when you modify this file.
 
-TUCCSA2::Application.config.session_store :cookie_store, key: '_TUCCSA2_session'
-
-# Use the database for sessions instead of the cookie-based default,
-# which shouldn't be used to store highly confidential information
-# (create the session table with "rails generate session_migration")
-# TUCCSA2::Application.config.session_store :active_record_store
+Rails.application.config.session_store :cookie_store, key: '_tuccsa2_session'
